requirements.txt
# Core dependencies
numpy>=1.21.0
opencv-python>=4.5.3
torch>=1.9.0
torchvision>=0.10.0

# YOLO dependencies
ultralytics>=8.0.0  # YOLOv8

# CVAT integration
cvat-sdk>=2.0.0

# Video processing
moviepy>=1.0.3
ffmpeg-python>=0.2.0

# Utilities
tqdm>=4.62.0
PyYAML>=6.0
python-dotenv>=0.19.0
requests>=2.26.0

# Development dependencies
pytest>=6.2.5
black>=21.9b0
flake8>=3.9.2 